Added DAC1 File Send option

+++ TargetEchoGen/FileTransferAgent.h
@@ -77,7 +77,7 @@ private:
     QString _sFilePath;
     QString _IPAddress;
     quint16 _Port;
-    int _iDataSize;
+    qint64 _iDataSize;
 
     QMutex mutex;
     bool abort;
